Renaming package to ‘terra-clinical-data-grid’ for consistency’s sake.
