+++ packages/terra-clinical-data-grid/LICENSE
