+++ packages/terra-clinical-data-grid/docs/README.md
@@ -1,11 +1,11 @@
-# Terra DataGrid
+# Terra Clinical DataGrid
 
-The DataGrid is an organizational component that renders a data collection in a grid-like format.
+The DataGrid is an organizational component that renders a collection of data in a grid-like format.
 
 ## Getting Started
 
 - Install with [npmjs](https://www.npmjs.com):
-  - `npm install terra-data-grid`
+  - `npm install terra-clinical-data-grid`
 
 ## Usage
 
@@ -24,8 +24,8 @@ Columns are provided to the DataGrid through two props: `pinnedColumns` and `ove
 |**id**|string|required|none|An identifier for the column. This identifier should be unique across the set of columns provided to the DataGrid.|
 |**width**|number|optional|none|An number (in px) specifying the width of the column. If not provided, the DataGrid's default column width will be used.|
 |**isSelectable**|bool|optional|none|A boolean indicating whether or not the column header is selectable. If true, the DataGrid's `onColumnSelect` prop will be called upon column header selection.|
-`|**isResizable**|bool|optional|none|A boolean indicating whether or not the column header is resizable. If true, a resize indicator will be rendered over the column header's contents, and the DataGrid's `onRequestColumnResize` function will be called upon interaction with the resize indicator. The new width will be provided with the function call so that the column's width property can be updated as needed.|
-`|**text**|string|optional|none|A string of text to render within the column header. This text will follow standard truncation rules as necessary. If a `component` value is specified, `text` will be ignored.|
+|**isResizable**|bool|optional|none|A boolean indicating whether or not the column header is resizable. If true, a resize indicator will be rendered over the column header's contents, and the DataGrid's `onRequestColumnResize` function will be called upon interaction with the resize indicator. The new width will be provided with the function call so that the column's width property can be updated as needed.|
+|**text**|string|optional|none|A string of text to render within the column header. This text will follow standard truncation rules as necessary. If a `component` value is specified, `text` will be ignored.|
 |**sortIndicator**|One of: **ascending**, **descending**|optional|none|A string indicating which sorting indicator should be rendered. If not provided, no sorting indicator will be rendered. If a `component` value is specified, `sortIndicator` will be ignored.
 |**component**|component|optional|none|A custom component to render within the column header. If provided, the `text` and `sortIndicator` values will be ignored.|
