+++ packages/terra-clinical-data-grid/package.json
@@ -1,26 +1,26 @@
 {
-  "name": "terra-data-grid",
+  "name": "terra-clinical-data-grid",
   "main": "lib/DataGrid.js",
   "version": "0.0.0",
   "description": "TODO",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/cerner/terra-framework.git"
+    "url": "git+https://github.com/cerner/terra-clinical.git"
   },
   "keywords": [
     "Cerner",
     "Terra",
-    "Framework",
-    "terra-data-grid",
+    "Clinical",
+    "terra-clinical-data-grid",
     "DataGrid",
     "UI"
   ],
   "author": "Cerner Corporation",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/cerner/terra-framework/issues"
+    "url": "https://github.com/cerner/terra-clinical/issues"
   },
-  "homepage": "https://github.com/cerner/terra-framework#readme",
+  "homepage": "https://github.com/cerner/terra-clinical#readme",
   "peerDependencies": {
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
@@ -34,18 +34,15 @@
     "classnames": "^2.2.5",
     "memoize-one": "^4.0.0",
     "prop-types": "^15.5.8",
+    "react-draggable": "^3.0.5",
     "react-lifecycles-compat": "^3.0.4",
+    "resize-observer-polyfill": "^1.4.1",
     "terra-base": "^3.0.0",
     "terra-button": "^2.0.1",
-    "terra-icon": "^2.6.0",
-    "terra-aggregator": "^2.4.0",
     "terra-content-container": "^2.0.1",
     "terra-doc-template": "^1.12.0",
-    "terra-overlay": "^2.16.0",
-    "terra-slide-panel-manager": "^2.3.0",
-    "terra-modal-manager": "^2.4.0",
-    "react-draggable": "^3.0.5",
-    "resize-observer-polyfill": "^1.4.1"
+    "terra-icon": "^2.6.0",
+    "terra-overlay": "^2.16.0"
   },
   "scripts": {
     "compile": "npm run compile:clean && npm run compile:build",
