+++ packages/terra-clinical-data-grid/src/DataGrid.module.scss
