+++ packages/terra-clinical-data-grid/src/proptypes/cellDataShape.js
