+++ packages/terra-clinical-data-grid/src/proptypes/columnDataShape.js
