+++ packages/terra-clinical-data-grid/src/proptypes/rowDataShape.js
