+++ packages/terra-clinical-data-grid/src/proptypes/sectionDataShape.js
