+++ packages/terra-clinical-data-grid/src/subcomponents/Cell.jsx
