+++ packages/terra-clinical-data-grid/src/subcomponents/HeaderCell.jsx
