+++ packages/terra-clinical-data-grid/src/subcomponents/HeaderCell.module.scss
