+++ packages/terra-clinical-data-grid/src/subcomponents/ResizeHandle.jsx
