+++ packages/terra-clinical-data-grid/src/subcomponents/ResizeHandle.module.scss
