+++ packages/terra-clinical-data-grid/src/subcomponents/Row.jsx
