+++ packages/terra-clinical-data-grid/src/subcomponents/Scrollbar.module.scss
