+++ packages/terra-clinical-data-grid/src/subcomponents/SectionHeader.module.scss
