+++ packages/terra-clinical-data-grid/src/terra-dev-site/doc/clinical-data-grid/About.doc.jsx
