+++ packages/terra-clinical-data-grid/src/terra-dev-site/doc/clinical-data-grid/DatagridExamples/DatagridWithColumnResizing.5.doc.jsx
